Fix sort order of modules in releng POM
Lexicographic order of our module lines looks strange when done on
complete lines. This is due to the fact that the closing tag is included
in the sort algorithm, which lead to shorter lines being included as if
they would contain module at the end. Thus no the sort is performed
before XML tags are added to the line.