Implement qualified version numbers in workflowHelper
We decided to use qualified version numbers throughout our build. These
qualifiers are resolved to v<date> for releases, to hotfix<Id>_v<date>
for hotfixes and to nightly_<date> for nightly builds. The former two
are controlled by the workflow helper, while the later one is the default
implemented within the parent POM.